Parents components was relocated to parents folder
