+++ src/app/parents/character-page/character-page.component.ts
