+++ src/app/parents/characters-page/characters-page.component.ts
@@ -1,11 +1,11 @@
 import { Component, OnInit, ViewChild } from '@angular/core';
 import { MatTable, MatPaginator, MatTableDataSource, MatDialog } from '@angular/material';
 
-import { Character } from '../shared/entities/character';
-import { CharactersEqualsHelper } from '../shared/helpers/equals-helpers/charactersEqualsHelper';
-import { CharactersService } from '../shared/services/characters.service';
-import { AddCharacterDialogComponent } from '../dialogs/add-character-dialog/add-character-dialog.component';
-import { RemoveCharacterDialogComponent } from '../dialogs/remove-character-dialog/remove-character-dialog.component';
+import { Character } from '../../shared/entities/character';
+import { CharactersEqualsHelper } from '../../shared/helpers/equals-helpers/charactersEqualsHelper';
+import { CharactersService } from '../../shared/services/characters.service';
+import { AddCharacterDialogComponent } from '../../dialogs/add-character-dialog/add-character-dialog.component';
+import { RemoveCharacterDialogComponent } from '../../dialogs/remove-character-dialog/remove-character-dialog.component';
 
 @Component({
   selector: 'app-characters-page',
